Bots automatically download new updates.
The C2 maintains a version ticker. Every time a change is registered, the version increments, and the impacted files are noted in a changelog table.

On each heartbeat, the C2 sends back the latest version number. If the bot notices its local version is outdated, it requests a list of the impacted filenames, and then downloads each of those directly. If successful, the new version number is saved.

WARNING: JANKY. Needs more reliable error handling, and the registration endpoint is clunky.